refactored functions into a different file updated .git ignore to add pycache
