+++ .gitignore
@@ -1,2 +1,3 @@
 # The books directory
 /books
+/__pycache__
